+++ boot/src/main/scala/com/stuckinadrawer/cookbook/recipes/Recipe.scala
@@ -4,7 +4,7 @@ import java.time.OffsetDateTime
 
 object Recipe {
 
-  final case class RecipeId(value: Long) extends AnyVal
+  final case class RecipeId(value: Int) extends AnyVal
 
   type Ingredient = String
